more testing done

+++ demo/src/test/java/excel/Test.java
@@ -297,11 +297,13 @@ public class Test {
 	                case 3:
 	                    System.out.println("Enter camp name to delete: ");
 	                    String newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
 	                    staff.deleteCamp(newCamp, database);
 	                    break;
 	                case 4:
 	                    System.out.println("Enter camp name to set visibility: ");
 	                    newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
 	                    staff.toggleCampVisibility(sc, newCamp);
 	                    break;
 	                case 5:
@@ -313,26 +315,31 @@ public class Test {
 	                case 7:
 	                    System.out.println("Enter camp name to view enquiries: ");
 	                    newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
 	                    staff.viewEnquiries(newCamp);
 	                    break;
 	                case 8:
 	                    System.out.println("Enter camp name to reply to enquiry: ");
 	                    newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
 	                    staff.replyEnquiries(sc, newCamp);
 	                    break;
 	                case 9:
 	                    System.out.println("Enter camp name to view suggestions: ");
 	                    newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
 	                    staff.viewSuggestions(newCamp);
 	                    break;
 	                case 10:
 	                    System.out.println("Enter camp name: ");
 	                    newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
 	                    staff.replySuggestions(sc, newCamp, database);
 	                    break;
 	                case 11:
 	                    System.out.println("Enter camp name to generate report: ");
 	                    newCamp = sc.nextLine();
+                        if(database.getCamp(newCamp) == null){System.out.println("No such camp"); break;}
                         System.out.println("How to print?");
 	                	System.out.println("1: Attendees only");
 	                	System.out.println("2: Camp committee only");
